Rename modules.graphql to modules.api #44
