+++ maverick_api/modules/api/maverick_status/__init__.py
@@ -20,8 +20,8 @@ from graphql import (
 )
 from graphql.pyutils.event_emitter import EventEmitter, EventEmitterAsyncIterator
 
-from modules.graphql import moduleBase
-from modules.graphql import schemaBase
+from modules.api import moduleBase
+from modules.api import schemaBase
 
 from modules.base.tornadoql.session_control import Session
