+++ maverick_api/modules/api/maverick_status/script.txt
